+++ AffiseAttributionLib/Classes/AffiseApi.swift
@@ -12,7 +12,8 @@ internal protocol AffiseApi {
     var initPropertiesStorage: InitPropertiesStorage {get}
     var preferencesUseCase: PreferencesUseCase {get}
     var postBackModelFactory: PostBackModelFactory {get}
-    var retrieveInstallReferrerUseCase: RetrieveInstallReferrerUseCase {get}
+    var retrieveReferrerUseCase: RetrieveReferrerUseCase {get}
+    var retrieveReferrerOnServerUseCase: RetrieveReferrerOnServerUseCase {get}
     var moduleManager: AffiseModuleManager {get}
     var storeInternalEventUseCase: StoreInternalEventUseCase {get}
     var debugValidateUseCase: DebugValidateUseCase {get}