+++ AffiseAttributionLib/Classes/referrer/RetrieveReferrerUseCase.swift
@@ -1,0 +1,31 @@
+import Foundation
+
+class RetrieveReferrerUseCase {
+
+    private var referrerURL: URL?
+    
+    /**
+     * Set referrerURL
+     */
+    func setReferrer(_ url: URL?) {
+        referrerURL = url
+    }
+    
+    func getReferrer() -> String? {
+        return referrerURL?.absoluteString
+    }
+    
+    /**
+     * Return referrer URL
+     */
+    func getReferrerUrl(_ callback: @escaping OnReferrerCallback) {
+        callback(referrerURL?.absoluteString)
+    }
+    
+    /**
+     * Return referrer URL parameter by key
+     */
+    func getReferrerUrlValue(_ key: ReferrerKey, _ callback: @escaping OnReferrerCallback) {
+        callback(referrerURL?.absoluteString.getQueryStringParameter(key.value()))
+    }
+}