+++ AffiseInternal/Classes/AffiseApiMethod.swift
@@ -26,8 +26,10 @@ public enum AffiseApiMethod: Int {
     case GET_PROVIDERS
     case IS_FIRST_RUN
     // callback
-    case GET_REFERRER_CALLBACK
-    case GET_REFERRER_VALUE_CALLBACK
+    case GET_REFERRER_URL_CALLBACK
+    case GET_REFERRER_URL_VALUE_CALLBACK
+    case GET_REFERRER_ON_SERVER_CALLBACK
+    case GET_REFERRER_ON_SERVER_VALUE_CALLBACK
     case REGISTER_DEEPLINK_CALLBACK
     case SKAD_REGISTER_ERROR_CALLBACK
     case SKAD_POSTBACK_ERROR_CALLBACK
@@ -73,8 +75,10 @@ public enum AffiseApiMethod: Int {
         case .GET_PROVIDERS: return "get_providers"
         case .IS_FIRST_RUN: return "is_first_run"
         // callback
-        case .GET_REFERRER_CALLBACK: return "get_referrer_callback"
-        case .GET_REFERRER_VALUE_CALLBACK: return "get_referrer_value_callback"
+        case .GET_REFERRER_URL_CALLBACK: return "get_referrer_url_callback"
+        case .GET_REFERRER_URL_VALUE_CALLBACK: return "get_referrer_url_value_callback"
+        case .GET_REFERRER_ON_SERVER_CALLBACK: return "get_referrer_on_server_callback"
+        case .GET_REFERRER_ON_SERVER_VALUE_CALLBACK: return "get_referrer_on_server_value_callback"
         case .REGISTER_DEEPLINK_CALLBACK: return "register_deeplink_callback"
         case .SKAD_REGISTER_ERROR_CALLBACK: return "skad_register_error_callback"
         case .SKAD_POSTBACK_ERROR_CALLBACK: return "skad_postback_error_callback"