+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [1.6.38] - 2024-08-06
+
+### Added
+
+-  Api `Affise.getReferrerUrl`.
+-  Api `Affise.getReferrerUrlValue`.
+-  Api `Affise.getReferrerOnServer`.
+-  Api `Affise.getReferrerOnServerValue`.
+
+### Removed
+
+-  Api `Affise.getReferrer`.
+-  Api `Affise.getReferrerValue`.
+
 ## [1.6.37] - 2024-08-05
 
 ### Added
@@ -86,6 +100,7 @@
 - Improve `AffiseModuleManager`
 - Update for demo app
   
+[1.6.38]: https://github.com/affise/sdk-ios/compare/1.6.37...1.6.38
 [1.6.37]: https://github.com/affise/sdk-ios/compare/1.6.36...1.6.37
 [1.6.36]: https://github.com/affise/sdk-ios/compare/1.6.35...1.6.36
 [1.6.35]: https://github.com/affise/sdk-ios/compare/1.6.34...1.6.35